feat: refactor wallet and token send components; improve structure and styling, and remove unused CSS files
