+++ apps/wso2-public-wallet/src/pages/confirm-token-send/confirm-token-send.css
