+++ apps/wso2-public-wallet/src/pages/send-tokens/send-tokens.css
